chore: use new dbus of dock daemon
becase dock daemon has been moved into dock, update new dbus path

log: as title
